+++ .idea/misc.xml
@@ -21,9 +21,9 @@
         <entry key="..\:/AndroidStudioProjects/Android-Suhyeon/app/src/main/res/layout/fragment_home.xml" value="0.2" />
         <entry key="..\:/AndroidStudioProjects/Android-Suhyeon/app/src/main/res/layout/fragment_image.xml" value="0.1" />
         <entry key="..\:/AndroidStudioProjects/Android-Suhyeon/app/src/main/res/layout/fragment_profile.xml" value="0.1" />
-        <entry key="..\:/AndroidStudioProjects/Android-Suhyeon/app/src/main/res/layout/fragment_repository.xml" value="0.25" />
+        <entry key="..\:/AndroidStudioProjects/Android-Suhyeon/app/src/main/res/layout/fragment_repository.xml" value="0.176" />
         <entry key="..\:/AndroidStudioProjects/Android-Suhyeon/app/src/main/res/layout/item_follower.xml" value="0.2" />
-        <entry key="..\:/AndroidStudioProjects/Android-Suhyeon/app/src/main/res/layout/item_repository.xml" value="0.25364583333333335" />
+        <entry key="..\:/AndroidStudioProjects/Android-Suhyeon/app/src/main/res/layout/item_repository.xml" value="0.176" />
         <entry key="..\:/AndroidStudioProjects/Android-Suhyeon/app/src/main/res/menu/menu_bottom.xml" value="0.1" />
       </map>
     </option>
