+++ app/src/main/AndroidManifest.xml
@@ -13,7 +13,7 @@
         android:theme="@style/Theme.SOPTHub"
         android:usesCleartextTraffic="true">
         <activity
-            android:name=".ui.view.SignInActivity"
+            android:name=".ui.view.user.SignInActivity"
             android:exported="true">
             <intent-filter>
                 <action android:name="android.intent.action.MAIN" />
@@ -21,9 +21,9 @@
                 <category android:name="android.intent.category.LAUNCHER" />
             </intent-filter>
         </activity>
-        <activity android:name=".ui.view.SignUpActivity" />
+        <activity android:name=".ui.view.user.SignUpActivity" />
         <activity android:name=".ui.view.MainActivity" />
-        <activity android:name=".ui.view.DetailActivity" />
+        <activity android:name=".ui.view.profile.DetailActivity" />
     </application>
 
 </manifest>
