+++ app/src/main/java/org/sopt/sopthub/ui/view/home/HomeFragment.kt
@@ -1,14 +1,14 @@
-package org.sopt.sopthub.ui.view
+package org.sopt.sopthub.ui.view.home
 
 import android.os.Bundle
 import android.view.LayoutInflater
 import android.view.View
 import android.view.ViewGroup
 import org.sopt.sopthub.R
-import org.sopt.sopthub.databinding.FragmentProfileBinding
+import org.sopt.sopthub.databinding.FragmentHomeBinding
 import org.sopt.sopthub.ui.base.BindingFragment
 
-class ProfileFragment : BindingFragment<FragmentProfileBinding>(R.layout.fragment_profile) {
+class HomeFragment : BindingFragment<FragmentHomeBinding>(R.layout.fragment_home) {
 
     override fun onCreateView(
         inflater: LayoutInflater,
