+++ app/src/main/res/layout/fragment_follower_list.xml
@@ -10,14 +10,12 @@
     <androidx.constraintlayout.widget.ConstraintLayout
         android:layout_width="match_parent"
         android:layout_height="match_parent"
-        tools:context=".ui.view.FollowerFragment">
+        tools:context=".ui.view.profile.FollowerListFragment">
 
         <androidx.recyclerview.widget.RecyclerView
             android:id="@+id/rv_follower"
             android:layout_width="match_parent"
             android:layout_height="match_parent"
-            android:layout_marginHorizontal="30dp"
-            android:layout_marginTop="20dp"
             app:layoutManager="androidx.recyclerview.widget.LinearLayoutManager"
             tools:listitem="@layout/item_follower" />
